requirements.txt
asgiref==3.5.2
autopep8==1.7.0
bcrypt==4.0.0
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
psycopg2-binary==2.9.3
pycodestyle==2.9.1
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
